Move lease auth tests
